requirements.txt
Django==4.2.7
Pillow==11.0.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-decouple==3.8
pytesseract==0.3.10
numpy==1.26.4
opencv-python==4.8.1.78
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
google-cloud-vision==3.11.0
openai>=1.12.0
google-generativeai>=0.8.0
google-genai>=0.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
